Consolidate override logic into model/WebglRenderer
