+++ addons/xterm-addon-webgl/src/GlyphRenderer.ts
@@ -191,47 +191,11 @@ export class GlyphRenderer {
       return;
     }
 
-    // Get any decoration foreground/background overrides
-    const decorations = this._decorationService.getDecorationsOnLine(y);
-    let bgOverride: number | undefined;
-    let fgOverride: number | undefined;
-    for (const d of decorations) {
-      const xmin = d.options.x ?? 0;
-      const xmax = xmin + (d.options.width ?? 1);
-      if (x >= xmin && x < xmax) {
-        if (d.backgroundColorRGB) {
-          bgOverride = d.backgroundColorRGB.rgba;
-        }
-        if (d.foregroundColorRGB) {
-          fgOverride = d.foregroundColorRGB.rgba;
-        }
-      }
-    }
-
-    // Convert any overrides from rgba to the fg/bg packed format. This resolves the inverse flag
-    // ahead of time in order to use the correct cache key
-    if (bgOverride !== undefined) {
-      // Non-RGB attributes from model + override + force RGB color mode
-      if (fg & FgFlags.INVERSE) {
-        bgOverride = (bg & ~Attributes.RGB_MASK) | (fgOverride !== undefined ? fgOverride >> 8 : fg) | Attributes.CM_RGB;
-      } else {
-        bgOverride = (bg & ~Attributes.RGB_MASK) | bgOverride >> 8 | Attributes.CM_RGB;
-      }
-    }
-    if (fgOverride !== undefined) {
-      // Non-RGB attributes from model + force disable inverse + override + force RGB color mode
-      if (fg & FgFlags.INVERSE) {
-        fgOverride = (fg & ~Attributes.RGB_MASK & ~FgFlags.INVERSE) | (bgOverride !== undefined ? bgOverride >> 8 : bg) | Attributes.CM_RGB;
-      } else {
-        fgOverride = (fg & ~Attributes.RGB_MASK & ~FgFlags.INVERSE) | fgOverride >> 8 | Attributes.CM_RGB;
-      }
-    }
-
     // Get the glyph
     if (chars && chars.length > 1) {
       rasterizedGlyph = this._atlas.getRasterizedGlyphCombinedChar(chars, bg, fg);
     } else {
-      rasterizedGlyph = this._atlas.getRasterizedGlyph(code, bgOverride ?? bg, fgOverride ?? fg);
+      rasterizedGlyph = this._atlas.getRasterizedGlyph(code, bg, fg);
     }
 
     // Fill empty if no glyph was found
